refactored twitter functions into package
